feat(SKILL-163): retain telemetry install_id across a disable
Telemetry disable now rewrites config.json in place with level off instead of
deleting the file, so install_id survives an off -> anonymous round trip.
delete() is gone from the TelemetryConfigStore port and FileTelemetryConfigStore,
making config destruction a compile error rather than a silent regression.
Keys not owned by the level mutation (external_addon_sources, execution_matrix)
are preserved by the single withTelemetryLevel domain rule.
