+++ .feature-specs/SKILL-163-telemetry-release-attribution/decomposition-manifest.yaml
@@ -9,18 +9,18 @@ base_branch: "main"
 feature_branch: "feat/SKILL-163-telemetry-release-attribution"
 stack_branches: []
 current_subtask_intent:
-  subtask_id: 1
+  subtask_id: 2
   action: "resume"
 subtasks:
 - id: 1
   name: "Release attribution on every telemetry event"
   spec_path: ".feature-specs/SKILL-163-telemetry-release-attribution/spec_subtask_1_release_attribution.md"
-  status: "pending"
+  status: "complete"
   branch: null
   commit_sha: null
   workflow_id: "wftr-20260806-074855-jqui"
   blocked_reason: null
-  last_resumable_step: null
+  last_resumable_step: "commit_push"
   linear_issue_id: null
   finalizing_agent_id: null
   participating_agent_ids: []
@@ -31,7 +31,7 @@ subtasks:
   status: "pending"
   branch: null
   commit_sha: null
-  workflow_id: null
+  workflow_id: "wftr-20260806-081019-b8uz"
   blocked_reason: null
   last_resumable_step: null
   linear_issue_id: null
